fifth

+++ app/Http/Controllers/OwnerController.php
@@ -10,13 +10,20 @@ use App\Models\Owner;
 class OwnerController extends Controller
 {
     public function index(){
+
+        if (Auth::user()->isAdmin() || Auth::user()->isReadingUser()) {
         $owners = Owner::with('cars')->get();
+        } else {
+            $owners = Owner::with('cars')->where('user_id', Auth::id())->get();
+        }
+
         return view('owners.index', ['owners' => $owners]);
     }
-    public function create(){
 
+    public function create(){
         return view('owners.create');
     }
+
     public function store(Request $request){
         
         $data = $request->validate([
@@ -31,11 +38,15 @@ class OwnerController extends Controller
             'address' => 'required'
         ]);
 
+        $data['user_id'] = Auth::user()->id;
+
         $newOwner = Owner::create($data);
 
         return redirect(route('owners.index'))->with('success', trans('Added'));
     }
     public function edit(Owner $owner){
+
+        $this->authorize('update', $owner);
 
         return view('owners.edit', ['owner' => $owner]);
     }
@@ -58,6 +69,8 @@ class OwnerController extends Controller
         return redirect(route('owners.index'))->with('success', trans('Updated'));
     }
     public function destroy(Owner $owner){
+
+        $this->authorize('delete', $owner);
 
         $owner->cars()->delete();
